Handle ShareJS connection's `disconnected` event
This event actually affects the `connection.state` which will be picked
up by the sharejs service's `stateChangeHandler`.

In the UI this leads to the textarea value being reset as soon as the
`disconnected` event is sent from the ShareJS connection which in turn
is notified of that by the underlying transport socket.

[#59115560]