Create $apply-aware Angular adapter
So the burden of calling $apply() lies entirely within the adapter
and other code does not have to bother with it.

This means also to manually call $apply() in specs manually after
calls to `yield`. Originally the callback triggered by yield would
call apply, now we have to do that ourselves.