Squash ooyala-widget commit history
This is the result of a rebase to squash all the commits
in the history of the ooyala widget. It includes all the work
from the basic widget til the multiple asset one.

At the time of this squash the multiple asset functionallity
works pretty well with only some performance issues when many
videos are loaded on one entry.