Check first if the Enter key has been pressed
Check first if the Enter key has been pressed so the event can
always (not only when the typeAhead is disabled) be prevented and the progation stopped.

The logic inside the function which handles the 'keyup' event remains (more or less) the
same:

Logic before the change:

- If the key pressed is Enter and the directive doesn't have typeAhead
  enabled then the model is immediately updated.
- If the key pressed is Enter and the directive has typeAhead enabled then
  a debounced update is used.

Logic after the change:

- If the key pressed is Enter it doesn't matter if the directive has or
  hasn't typeAhead enabled. The model is immediately updated.
- If the key pressed is not Enter then the same rules before apply. If
  the directive has typeAhead enabled then a debounced update happens
  otherwise the directive does nothing.

The only difference is the delay in calling the update method.

But applying this change allows to use the cf-search directive (with typeAhead enabled)
inside the modal dialog without having the dialog closed because of pressing the
Enter key.