Don't apply the ng-show to the wrapper over the contents
The ng-show directive applied to the element that wraps the error
messages and the players yields useless the transitions applied to the
player directive.

Instead of hiding/showing it we add/remove the marging (the current goal of
the wrapper) depending on the existence or not of content inside the wrapper.