Provide information on uniqueness of a slug.
The space is queried for entries other than the current one that
have the same slug text, but are published. If no such entry exists,
the slug is deemed unique.

This adds a status icon to the slug widget's input field, which
displays one of three statuses: Unique, Duplicate, or Checking.

Also adds the appropriate tests.