Add a new google maps widget, and split the location editor.
- This splits the google maps widget code from the location editor,
and turns both into independent location editing widgets.
- Modernifies the unit tests of the location eidtor according to the new
  conventions.

This means that now the google maps widget and the location editor
widget can now be used without/instead of each other as they provide the
same collaborative location picking functionality and operate on the
same `location` model. Both widgets are independently tested.

The built-in location editor tool now places both these widgets
in the same scope via the template, which allows them to be linked
to each other and still function in a sharejs-friendly manner.

Ultimately, the mixed tool allows one to edit the location even while
waiting for the google maps API to be loading.