Increase clickable area for checkboxes in dropdown
This concerns the “View” dropdown in the Entry Editor. Before the
clickable `label` was rendered as an inline element and did not fill
the whole row in the dropdown. This was particularly confusing since
the visual hints for interacting (pointer cursor and hover accent
color) where presented on the `li` block element.