Diable totango and bugsnag when loading fails
If the user blocks loading of external scripts, don't leave the
CallBuffer open, creating memory leaks.

To write the tests I had to sin and stub a method on the internal
_buffer property of bugsnag and totango. The Alternative would have been
to create a complicated Alternative stub for the CallBuffer service.

But I'm absolved, there's a guy from google saying it's ok:
http://philipwalton.com/articles/how-to-unit-test-private-functions-in-javascript/

Also, this is a very hard case to test: Per definition disabling the
buffer is externally indistinguishable from buffering the calls.
The effect (memory leak) is purely behind the scenes.