Reorder elements in space-view
Move account iframe below (in the file) the advices, so that it appears
on top of them (z-index) in the app. This fixes the problem where a user
with no spaces was not be able to access the account settings because
the "You have no spaces" advice was obstructing the iframe.