Update entry link autocomplete if validation changes
The link editor controller uses the field’s `linkContentType`
validation to display only the valid content types in the autocomplete.
Before the filter was only setup when the controller was instantiated.
This meant that updating the content type validation and returning to
the entry editor would not update the filter. Now we watch the
validation property to automatically update the filter when the
validation changes.

See [TP#3399](https://contentful.tpondemand.com/entity/3399).