Replace field settings list
Removed the list of field settings editor and replaced it with simple
items and a dummy link to the field settings dialog. In particular we
removed the `cf-field-settings-editor` directive and the corresponding
controller.

Also removed the `cf-content-type-field-list` directive. Its tasks are
now handled by a custom `cf-ui-sortbale` directive.

The template for a single field is now merged into the
`content_type_field_list` template and radiacally simplified. These
changes are also reflected in the removal of a lot of styles.