Fix logging server errors
In the past we logged errors in a different way, which required us to
specify an error property in the object passed to the logger. This is
not really required anymore, and the code written to flatten errors for
better visibility in Bugsnag was expecting this to be the case, and
therefore not working.