Updated trial subscription detection
With Zuora, the organization.subscriptionState will never be "trial"
during trial subscriptions. Instead it is set to "pending", which
will no longer indicate that the organization is using a trial.
With Zuora, the organization.trialPeriodEndsAt property will always
be null if the organization is not in a trial, so this is now being
used for the determination.