Makes ui expect non-Zuora subscription data
The Gatekeeper provided token looks different depending on whether
Zuora or the previous subscription system is in place for the
current organization. Maintaining both systems with two different
token formats would be cumbersome in the ui. Therefore the ui
is - again - only expecting the legacy system for now.

The token format will probably be unified for both systems in the
future. After that, the paywall will also be activated again.