Reactivates Paywall
The Gatekeeper token's format has changed to look the same for
Organizations using Zuora and lagacy in contentful/gatekeeper#712.

As a consequence, this change is reactivating the paywall again.

For non-Zuora Organization users the paywall should not be shown.
This handling is not required in the UI though since for those
users the subscription will change to the free limited one
automatically as soon as the trial ends. This also makes the
existing tests for this case irrelevant, which get removed with
this change.