Fixes markdown fullscreen editing
Markdown fullscreen mode has been broken since 49b8460, as some
elements got displayed on top of the editor. The reason for
this was that the change above has introduced an explicit z-index
for the workbench, which made it the new stacking context for the
.cf-zenmode fullscreen markdown editor. At the same time, the top
menu bar's and breadcrumb's z-indexes were higher since they had
to be displayed on top of the workbench of course.
This change sets the workbench's z-index back to auto and there-
fore destorys the undesired stacking context.