refactor(widgets): Remove static widgets
We can remove static widgets because they were never rendered.

Static widgets are marked by the `widgetType: 'static'` property  in the editing
interface data returned by the server. They do not have a `fieldId` property.
Since the `syncOrder()` function in the `editingInterfaces` service filters all
widgets that are not attached to a content type field, static widgets are never
included in the editing interfaces returned by that service. Hence, they are not
rendered and we can remove them safely.

Note that this commit removes a failing test case which asserted the presence of
static widgets which is further evidence that static widgets are unused.