refactor(widgets): Simplify building widget list from content type fields
The purpose of `syncWidgets()` is to provide an array of widgets that
corresponds to the content type field array. Before, this was
accomplished by removing widgets that did not map to a field, adding
widgets for fields that had no widgets, and then rearraning the order
of the widgets to match that of the fields.

We can accomplish all this in a single function.