refactor(widget-render): Move locale computation to cf-form-widget
Before, the top-level entry editor directive used the
'FormWidgetsController' to build a list of renderable widgets the
included a 'locales' property. The 'locales' list was then iterated
over in the template to create a widget for each locale.

Now we separate the generation of the 'locales' list by moving it some
levels deeper into the 'cfFormWidget' directive. This allows us to
ignore the locales in the 'widgets' service.