Fixes widgets' use of Channel.send('connect', ...)
The params given to `send()` had to be an array while an object was
given. In the widget-sdk this has never been an issue since the
`Channel` message handling has not been used to handle 'connect'.
With https://github.com/contentful/widget-sdk/pull/71 this will
change and params is required to be an array to be passed on to
the the widget-sdk Channel's message handlers.