Fixes ng-repeat crashes in entry editor
This is a hot-fix for https://contentful.zendesk.com/agent/tickets/5415

Tracking the `ng-repeat` by `widget.id` is not feasible anymore since
change eec3676 resulted in duplicate ids in some rare cases as for the
zendesk ticket's customer above.

Example of a Content Type's `.fields` causing this problem:
```js
[{
  "name": "Address",
  "id": "abcdef12345",
  "type": "Symbol",
  "apiName": "ADDRESS"
},
{
  "name": "Address",
  "id": "ADDRESS",
  "type": "Text",
  "disabled": true
}]
```
For the first field the `apiName` would be taken as the editingInterface's
respecitve `widget.id` while for the second field the `id` would be taken
(due to lack of an `apiName`). This will result in two `widget.id`s based
on the same string, `"address"`.