fix(navigator): Handle redirects before confirmation
Redirects are the first thing that is handled by the state change
listener. If a redirect occurs, we do not proccess the state change
further.

This prevents asking for confirmation twice. Before, on a state change
the user was asked to confirm, then the redirect happened so the user
was asked to confirm again.

[TP#7192](https://contentful.tpondemand.com/entity/7192)