Disables file upload image cropping step
Calls filepicker without the "CONVERT" service which enables the
image cropping on uploads.
Unfortunately we can not just take the objekt keys of
`filepicker.services` and remove "CONVERT" since that list would
also include "SEND_EMAIL", "GDRIVE", "CUSTOMSOURCE", "GENERIC"
and "AUDIO" which cause the filepicker to crash. The library seems
to lack a viable way to just exclude a certain service. Instead we
provide the whole list of services we want to enable.