fix(ct-editor): Set state to not dirty before redirecting to new CT
Fixes a regression introduced in e4435327ad62019488b6f9150360a217f71d2a18.

When we redirect after a new CT is created the state change handlers
tests for the `context.dirty` flag. Previously this was implicitly
handled by a watcher on the CT forms `$dirty` property. But that
watcher failed to trigger before the redirection took place.