fix(video-widgets): Prevent entry from being updated on open
We only update the ShareJS document when the video id has changed. This
prevents an increase in the entries version number when opening the
entry.

We also do not watch the attribute `value={{fieldData.value}}` but
`$scope.fieldData.value` directly. This prevents an update with the
uninterpolated string value `{{fieldData.value}}` when the directive is
rendered initially.

[TP#8403](https://contentful.tpondemand.com/entity/8403)