fix: Space templates doe not convert EI data
There is no reason for the space template creator to parse and change
the editing interfaces that are part of the template space. In fact
this broke due to the new schema we introduced.

Now we pass the editing interface data we get from the template space
on to the target space since we can assume that anything the server
returns is valid.