refactor: Update entity data from ShareJS centralized
Whenever the ShareJS document changes the data of the entity object on
the scope needs to be updated, too.

Previously this was accomplished by calling `otUpdateEntityData()`
after the document was opened and on each remote operation. For
locale changes we called `$scope.otDoc.updateEntityData()` at various
places in the app.

Now we update the entity data when the “change” event is fired. Since
this happens on local changes and remote operations we have covered all
bases.