fix(asset-list): Do not reload assets twice
Throttled list refresh was called twice before. Once after creating the
assets and a second time after processing them. We do not need to wait
for the processing since that has nothing to do with the presence of
the files in the list. We also remove the throtteling because the reset
is only called when assets have been uploaded which is done manually.