fix(entry-editor): Do not add or remove links twice
The `scope.links` value is used to render the link list. It is synced
through `ngModel` with `fieldData.value`, which in turn is linked to the
actual field data in `scope.entity.data`.

When updating through ShareJS the entity data is now updated
automatically, so there is no need to call `updateModel()` to write back
the modified link list to `scope.entity.data`. We will receive it
automatically from `fieldData.value`.