refactor: Do not inject config into source code
Instead of modifying the `src/javascripts/environment.js` file on the
fly to inject the current git revision we load it from
`window.CF_UI_VERSION`. This value is set when the index file is
deployed (see `bin/process_hosts`).

We also disable the version reloading in the development environment
and refactor the `environment` service.