Revert `cfnext-btn-...` styles using flexbox
`display: flex` does not work for `button` in Firefox [1]. A workaround
would be to have a `span` container inside the button which would use
flexbox. This would add additional overhead to the markup, especially in
the widget-sdk. Making the button a directive could act as a facade for
such more verbose markup but it wouldn't solve it for the widget-sdk.

1: https://bugzilla.mozilla.org/show_bug.cgi?id=984869