feat: Introduce streams to document editing
We removed the 'otChange' and 'otDocReady' events emitted in the
'Document' controller and replaced them with a 'changes' stream. Both
events were used by the 'DataFields' constructor and the
'cfIframeWidget' directives. This is migrated to use the 'changes'
stream and its derivative 'valueStreamAt'.

We also use 'valueStreamAt' in the 'otPath' directive. It replaces
the listeners for 'otRemoteOp' and 'otValueReverted'. Here we need
to filter for values that are set through the directives 'set()'
method. This allows us to get rid of those events, too.