Clean up new links editor icons, use currentColor
Improves the SVG markup to use `currentColor` and transparent
background plus filters instead of using `white` and assuming the
background is white as well.
I've done this for fun as a side project at home to get a little
more familiar with SVG and think it's much cleaner than the initial
markup generated by some graphics program.