Failing security audit does not make builds fail
Since we added the feature to check for vulnerabilities the process has
been like this: A new vulnerability is added to NSP. Our build fails.
We add an exception while basically all builds are blocked. We merge
master back to all the branches. We wait until a fix is released and
propagated up the packages. This takes super long.

Making our builds fail when the audit fails thus offers no benefit.