fix: Clone snapshot values to prevent invalid state
Before the following would happen:

  var o = doc.getValueAt(['a'])
  o.val = true;
  doc.setValueAt(['a'], o) // => no change is emitted

ShareJS would never send an update to the server. The reason for this
is that `setValueAt()` delegates to `ShareJS.setDeep()` which compares
the new value with the old one, stored in the internal `snapshot`
property. If they are equal no update is triggered.

Since `getValueAt()` would return a _reference_ to the object in the
`snapshot` property we modified the object inplace and the comparison
always succeeds. The `setDeep()` function think nothing has changed and
does not trigger an update.

[TP#12292](https://contentful.tpondemand.com/entity/12292)