Fix dropdown button with icon in firefox
With the `.x--block` buttons icon possitioning fix introduced in 52d7929
a issue with dropdown buttons with icon was introduced in Firefox.
This is due to a Firefox floating bug
(see https://bugzilla.mozilla.org/show_bug.cgi?id=488725).

There two simple solutions:
- Use `white-space: noraml`, not `white-space: nowrap` on `button`
- Position `cf-icon.btn-dropdown-icon` before `cf-icon.btn-icon` nodes

This change uses the first solution as it seems that `nowrap` is not
necessary for buttons other than `.btn-action-select__selector` buttons.