Fix crash caused by phantom user
Bugsnag: https://app.bugsnag.com/contentful/user-interface/errors/5808be8b1aeffcbdf8e03f9c?event_id=5808be8cf1322f11005402c6

So, if an entry was created by a user that was then removed from
the space, it would cause the app to crash as no user would be
returned for the id of the user we provide. This is because the
user list we maintain is scoped to the space and since the user
isn't a member of the space anymore, his/her id is not a part of
the user list for that space.

This fix mimics content list behaviour and defaults to an empty
string as the user name if no user is found.