refactor: Remove 'otDoc.state.error' property
This commit gets rid of the `otDoc.state.error` property and replaces it
with the `otDoc.status$` kefir property.

To accomplish this we refctor the DocLoader object to also emit errors
that come from the ShareJS connection being disconnected. This allows us
to remove the `error` kefir property from the DocConnection object.

Now in the entity editor Document we simply pass the doc loader error
state to the DocumentStatus function to create the `status$` stream.
This stream contains the business logic from the
StatusNotificationsController. That controller is in turn replaced by
the thin `cfStatusNotification` directive.