fix: Repaint MD editor when being displayed
Fixes [TP#12399](https://contentful.tpondemand.com/entity/12399)

Both the embedded Google Map and the CodeMirror instance must be
repainted whenever the element that contains the becomes visible. To
achieve this we introduce the 'ui/ResizeDetector' module that is based
on the [`element-resize-detector` library][erd-lib]. This module allows
us to observe size changes of a DOM element.

[erd-lib]: https://github.com/wnr/element-resize-detector