Set limit on query for linked entities
We need to set the limit on the query for linked entities to match the
number of entities in the request. Previously up to 200 entities were
requested at a time but the default limit of 100 was being applied so up to
half of the entities might not be loaded.

This came up in https://contentful.zendesk.com/agent/tickets/13555