Clean up entry list search logic and fix bug#14056
Clean up description:
There was unnecessary logic wrt the old endless-container which
isn't used anymore. All of that was plucked out and the code should
now be a bit more straightforward.

The bug had to do with bulk actions causing incorrect page to be shown.

Bug description:
When User selects all the the Entries on Content page and clicks on
“Delete”, “You have no entries yet” message is displayed, even only 40
were deleted and more entries exist.