refactor: Only refresh published CTs after space creation
The purpose of the handler in the 'CreateSpace' module was to update
the content types used by the learn view. Since that view only cares
about _published_ CTs we only refresh those.

Also removes the 'reloadEntries' event handler from the entry list.
This does not have an impact since the entry list is not shown when the
space creation finishes.