fix: Don’t crash entry preview component when entry is deleted
The problem is the `entryDocument.getValuePropertyAt([])` is set to
`undefined` when the entity is deleted by another user while the entry
editor is open. This crashes the app.

We expose the `data$` property that properly represents the data
property.

Bugsnag: https://app.bugsnag.com/contentful/user-interface/errors/589b676baa49e1c67625dbb9