fix: Re-enable entry editor when reconnecting
When the ShareJS connection drops we call the `close()` method. At this
point the ShareJS document is in state `open` and registers the close
callback. Then the connection tells the document it was disconnected
and the document emits the `closed`. But it never calls the registered
callback.

This means that the `close()` promise is never resolved. And in turn
`waitAndOpen()` is never resolved. By listening to `closed` we fix this.

[TP#14640](https://contentful.tpondemand.com/entity/14640), [TP#14369](https://contentful.tpondemand.com/entity/14369)