fix: Infinite loop when state resolution fails.
If space resources always respond with a 429 the following happens:
We go to any page of the app and the space state is activated. This
fails because it makes a request to the space’s content types. The
state change error handler kicks in and redirects the user to the
`home` state. The `home` state in turn redirects the user to the last
used space state. Here we begin anew