refactor: Simplify context history
The purpose of the `addToContext` parameter is to control whether a page
we navigate to should be added to the stack of pages we can go back to.
We only want this behavior for entries and assets.

That is why we reset the `contextHistory` for all other states. This is
accomplished by removing the `addToContext` parameter and using
`contextHistory.set()`.

Also not that we removed adding to the context history for list states.
These are top level states. And the context history would only contain
one element, the list itself. The breadcrumbs directive would thus not
show up.