fix: Index alias for SystemJS modules in tests
Before the following would happen. If a module for the file
`a/b/index.js` was registered we created a SystemJS map entry for 'a/b'
to 'a/b/index'. With the way SystemJS resolves modules this would lead
to `a/b/x` being resolved to `a/b/index/x`. The new approach fixes this.