Remove 'spaceContext.contentTypes'
This property was used in the content type list and the content type
metadata dialog.

For the list we use `space.getContentTypes()` directly instead of
calling `spaceContext.refreshContentTypes()` and then get the result
through `spaceContext.contentTypes` implicitly.

For the metadata dialog we pass the content type IDs in explicitly.
They are loaded in the content type editor controller.