Initial implementations for 'utils/Collections'
The module will hold functional code for dealing with frozen JavaScript
objects. It only contains a couple of methods for now but will by extended in the
future.

This approach is an alternative to using libraries like Immutable and
Mori. The big advantage is that we can use these functions without
changing the underlying data layer.

At first I considered using [icepick][icepick]. But unfortunately this
library is distributed as ES6 and will break our build system.

[icepick]: https://github.com/aearly/icepick