Do not remove default directory
Before we would remove the default view folder through
`cleanDefaultFolder()` when the last view from the folder was removed or
moved to another folder.

If a new view was created (`addViewToDefault()`) we would also ensure
that the default folder exists (`createDefaultFolder()`) and add the
view to it.

If the default folder did not exist but the user was dragging a view we
would show a temporary default folder as a drop target. This never
worked though, a user could never drop a view into the temporary default
folder.

We remove all this logic and simply make sure that a default folder
always exists.