fix: Normalize OtDoc snapshot wheneve remote op is received
With a recent update in the backend we might receive remote operations
that update the whole snapshot of a document. This means that deleted
fields and locales stored in ShareJS are pushed to the UI again.
Previously this only happened when the document was opened.

To solve this issue we now normalize the document whenever the whole
snapshot or the `fields` property change. For this we listen to the
`change` event stream. Since a change event is emitted when the document
is opened this also covers the previous use case.