fix: Throw to generator from Angular promises
If you throw an error in an Angular promise `then()` handler this error
cannot be caught by the next `catch()` block. Instead it is thrown to
`scope.$apply()`.

For this to work with generators we need to catch the errors thrown when
calling `$apply()` and throw these to the generator.