Fix crash on space creation failure
If a user does not yet have a space and tries to create on but the
creation fails the app crashes. The reason behind this is that the space
context is never initialized since there is no space. This leads to
`spaceContext.organizationContext.organization` crashing. We work around
this by using the organization that we wanted to create the space on.

See https://app.bugsnag.com/contentful/user-interface/errors/59ad233bc927570019844fd3