Redirects in states can be relative
Now the `redirectTo` property for states may have a relative state name.
A relative state is resolved with respect to the state that defines the
redirect.

This allows us to simplify the content type state definition. It also
fixes a bug where with breadcrumbs. The `navigation/CrumbFactory` module
that is used to generate links to views defines the state for content
types to be `...content_types.detail`. But this is an abstract state.
Clicking on a link for this will result in an error. See [1]

[1]: https://app.bugsnag.com/contentful/user-interface/errors/59cbc9e2d68b770018a082ee