Avoid digest loop errors in test generators
We lift $q promises in test runners so that `this.$apply()` does not
throw after yielding a $q promise.

Before, this code would throw a “Digest already in progress” error.

~~~js
it('calls $apply', function* () {
  yield $q.resolve();
  this.$apply();
})
~~~

If we do not lift $q Promises the statement `this.$apply()` will be
called from the `then` handler of the promise and thus in an Angular
digest loop. Calling `$apply()` in a digest loop throws an exception.