Add simple table styles
We implement the visuals for `table.table` with simplified CSS. The new
version does not split the table header and body. This gives us better
result when automatically laying out the column widths.

The new styles are used in the CMA Tokens list and the extensions list.
The look has not changed.