Track unexpected ui_config view migration errors (#2331)
Track unexpected ui_config view migration errors

We are not aware of any `searchTerm` that can break the migration code
but if the conversion from `searchTerm` to `searchText` and
`searchFilters` should fail unexpectedly, we now track it to bugsnag and
show an empty search for these views. The original `searchTerm` will
still be available as `_legacySearchTerm`.