Provide smaller interface for reactive storage
Instead of exposing a property bus we only expose the property. We still
can accomplish everything we need at the moment and this prevents
consumers from modifying the property.

We also completely rewrite the property creation to make sure we clean
up the event listeners correctly. Now `removeEventListener` is called
automatically when we unsubscribe from the property

We can also remove the call to `tokenMVar.empty()` since we call
`tokenMVar.put()` immediately after that.