Throw away access token on gatekeeper login
When the app loads following a redirect from a gatekeeper login we want
to throw away any existing token and get a new one from gatekeeper. This
ensures that the access token in local storage and the gatekeeper
session belong to the same user.

This fixes [TP#25314](https://contentful.tpondemand.com/entity/25314)
which describes the bug in more details.