Add codemirror enable/disable set/getFocus
This allows temporary disabling of the editor. Since in a disabled state
selecting text is still possible, the focus might be at a different
position when enabling. Therefore this also adds functions to get/set
the focus to be able to restore the focus at the same position after
enabling.