Use personal access tokens for local development
By opening the development version of the app with `#access_token=TOKEN`
we can circumvent gatekeeeper authenticatoin while developing locally.
It should be possible to use a personal access token as `TOKEN`.
Personal access tokens start with `CFPAT-`. Before, this was not matched by the
regexp that extracts the token from the hash.