Make 'ui/ResizeDetector' tests more stable
The 'ui/ResizeDetector' tests are flaky. At times we were unable to
detect a resize after waiting some time. Instead of waiting a fixed time
we now wait until we detect the change. If we never detect the change
the tests will fail with a timeout.

We also simplify the setup of the test cases.