Add root index and revision files while keeping package based deployment
This is a temporary change to test if things work correctly with
the new nginx config and the updated chef deploy script. Keeping
the package based deployment gives us a simple way of reverting our
changes without needing a build on the user interface side.

Once proven the new configs work, #2451 will be merged to retire
package based deployment.