Improve entity selector loading spinner
In c068269 we removed the entity selector's result list's loading
spinner below the search input. Instead we would show a spinner within
the search input.
The visual feedback when scrolling down was not great that way, so this
change improves upon this by showing the spinner below the list when
reaching the end of the list by scrolling. When searching, the spinner
within the search input is used.