allow linking assets of any locales in markdown editor
Before, if you try to link existing assets in the markdown editor,
and assets did not exist for the locale of the editor, we insert
nothing. However, user can omit assets localization, or just fine
with default locale image. This commit adds a warning to user,
which explains which assets will use which locale's version