Use cashed CTs in reference editor
This avoids sending multiple requests to the API with every
reference editor instance to avoid 429 HTTP errors.

Refactor conditions to render reference fields

Remove on-remove handler from bulk editor

This removes the added on-remove event handler to the bulk editor
and instead monkey patches the entityContext to add the remove handler.
This also improves the way we decide how to render the reference field
(inline, link or asset).