Take environment into account when closing state
Breadcrumbs currently do not record the environment correctly. This
means that the we go to the master environment state whenever
`closeState` is called.

We fix this by using the same code as in the breadcrumb directive that
dynamically uses the current state to determine whether to go back to an
environment.

[TP#27834](https://contentful.tpondemand.com/entity/27834)