Isolate scope for side panel trigger
The side panel trigger directive did not use isolated scope and messed
with the same scope object that is used by all the states. To be able to
isolate the scope we first need to lift the communication between the
trigger and the sidepanel to the enclosing controller.

This is a follow-up for #2721