Track entities created by us on the behalf of the user
Summary:
--------
As a PM I want to able to distinguish example content from “real” user
content so that I can say when the user has moved on to the “build”
phase of their time with Contentful.

The goal is to use this data to classify api calls as engaged api
calls.

TP: https://contentful.tpondemand.com/entity/27306-add-context-to-entity-create-operations