Deploy 06/04/2018 (#2788)
* Only set slug from title when field is enabled

We need to ensure that `setValue` for a slug field is only called when a
the user has permission to edit the field. We use the `isDisabled` flag
to determine that.

The reason for this is that if we call `setValue` for a field we are not
allowed to edit this operation might be bundled with other ops. ShareJS
will then reject _all_ operations because one of them is forbidden

See [ZD#25472](https://contentful.zendesk.com/agent/tickets/25472)

* Better fake for widget API

We change the implementation of the widget API mock so that it is closer
to the real thing. This forces us to update some tests.

* convert title update in slug editor to streams