Do not switch environments when visiting space state
We want to avoid that the web app switches to the master environment
when the user visits any of the space level states like the API key list.
To this end we provide all space level states also under the environment
state. Although this is conceptually wrong it allows us to retain the
current environment and prevent user errors.

[TP#28252](https://contentful.tpondemand.com/entity/28252)