Improve doc error handling
We adjust the `withRawDoc` function so that it includes the
`refreshAuth()` call. This reduces duplication in the `withRawDoc` calls
and makes the code less generic and more cohesive.

We also make sure that the local change event is not emitted multiple
times for `pushValueAt`

We also improve the test coverage by requiring the correct behavior for
`pushValueAt`.