Drop support for lc `slideIn` parameter (slidein)
Since we do not support lower case versions of other parameters (e.g.
`searchText`) we do not require to support it in this case either. Keep
support for lc "entry:" and "asset:" prefixes.
Moves some code of `EntityNavigationHelpers` to the bottom to
consistently have exports on top and helpers at the bottom of the file.