Add code to show the correct section in space home
This doesn't work yet as the code to update the value of all the flags
doesn't run when updated values are stored to localStorage.
Need to sort this out as the flag updation routine in
cf_onboarding_steps_directive needs to be called _after_ the space
creating and loading as completed as that is when it serialize the
space id to localStorage and that's when we should update the flags
that decide which next steps section to show.