Reapply pending ops when connection re-appears
If the ShareJS connection is disconnected there might be still pending
operations for a document that have not been acknowledged by the server.
Since the document is automatically reopened again after the connection
is reestablished we want to make sure that the operations are sent
again to the server and locally. We need to apply them locally too
because op reopening the doc we get reset the snapshot to the _server_
verions.