Prevents unncessary saves in entries with ST field (#3255)
When an entry has a structured text field with an embedded
entry, Slatejs sends a "set_node" operation that we don't
filter out and therefore triggers an autosave. This means when
a user enters a published entry, it changes its status from
Published to Pending Changes. So this adds "set_node" to the
list of operations to ignore.