Fix edit inline link operations being ignored
Throw an error if we encounter an operation with properties that we did
not previously expect when deciding whether or not the operation should
trigger a save.
This will allow us to quickly find bugs when developing new features
where we introduce new operation types or update properties we did not
previously touch.