[AUTH-484] Track markdown paste before/after size
Introduces tracking for paste events in the markdown field. We track the
text length of the field before and after the paste event. This allows us
to compare paste events in the structured text editor.

Our hypothesis is that editors will paste less often into the structured
text field, indicating that they are using the structured text interface
as a first line of recourse, rather than working elsewhere (e.g. in Google
docs).