Fix entity selector with single CT
A content type selector with only a single content type did not fetch
entries of only that one CT but all CTs since 62eff49. Instead of
fetching that single CT in the config, we now do this in the entity
selector itself as the config should not have to know about this at all.