[AUTH-429] Add rich text validations to CT UI
\## Purpose

Adds validations to the UI for the rich text content type:

- `size` (character count)
- Node validations:
  - Hyperlinks
    - `linkContentType`
  - Embedded block entries
    - `size` (# allowed)
    - `linkContentType`
  - Embedded inline entries
    - `size` (# allowed)
    - `linkContentType`

Removes some unnecessary validation boilerplate from the validation
decorator, added in AUTH-514 / #3388.

Cleans up a few utility functions for field validations.

\## Dependencies & References

**JIRA ticket:** https://contentful.atlassian.net/browse/AUTH-429

**Follows:**
- AUTH-514 (https://github.com/contentful/user_interface/pull/3388)
- AUTH-524 (https://github.com/contentful/contentful-validation/pull/114)

\## Approach

\### High level summary

Tweaks the field dialog logic to support UI for our link validations. Since
this is done programmatically, we end up reusing a lot of the UI logic.
Most of the remaining work involves adding support for node validation
serialization, which behaves basically the same way as field validations
(decorate the API-supplied validations for the UI -> update user-defined
validations in `$scope` -> undecorate and serialize everything on save).

\### Implementation details

Replicates much of the field validation decoration logic between the field
dialog controller and the validation decorator. The basic flow is the same:
we add a `nodeValidations` property to the field dialog controller's
`$scope` object and pass it to the validation decorator. The decorator
adds properties which are reflected in the content type validation UI. On
save, we extract the node validations in a serialized form compatible with
our API expectations, then add it as a `nodes` validation to our field
validations array (basically the same thing we're already doing, but with
some added extraction logic to support the slightly more complicated `nodes`
validation data structure).

\## TODOs

Either the validation decorator or the field dialog controller _really_
need some tests. Maybe as follow-up when we Reactify this area of the
codebase.